+++ libnopegl/src/text.c
@@ -178,8 +178,8 @@ static void set_geometry_data(struct text *s, struct text_data_pointers ptrs)
 
     /* Text/Box ratio */
     const struct ngli_box box = s->config.box;
-    const int32_t ar[] = {ctx->viewport.width, ctx->viewport.height};
-    const float box_ratio = (float)ar[0] * box.w / ((float)ar[1] * box.h);
+    const float ar[] = {ctx->viewport.width, ctx->viewport.height};
+    const float box_ratio = ar[0] * box.w / (ar[1] * box.h);
     const float text_ratio = (float)s->width / (float)s->height;
 
     /* Apply aspect ratio and font scaling */
